iOSウィジェットの状態判定を直読みに変更
